Initial commmit
Basic example of event propagation. In a structure like div1 > div2 > div3, if an event occurs in div3, if defined in div 3, it will be captured and later it will propagate to div2, div1 respectively. This is example of bubble phase. This is to show it is handled automatically by the js engine.